coulomb-temperature subpackage
