+++ coulomb/src/main/scala/coulomb/package.scala
@@ -35,26 +35,10 @@ package object coulomb {
    */
   type WithUnit[N, U] = Quantity[N, U]
 
-  /**
-   * An "infix" type alias for [[Temperature]]
-   * @tparam N The numeric representation type of the temperature value
-   * @tparam U The unit type of the temperature
-   * {{{
-   * import com.manyangled.coulomb._
-   * import SIBaseUnits._
-   * def f(t: Double WithTemperature Kelvin) = t + 100D.withUnit[Kelvin]
-   * }}}
-   */
-  type WithTemperature[N, U] = Temperature[N, U]
-
   /** enhances numeric types with utility methods for `coulomb` */
   implicit class CoulombExtendWithUnits[N](val v: N) extends AnyVal with Serializable {
     /** create a new unit Quantity of type U with the value of `this` */
     def withUnit[U]: Quantity[N, U] = new Quantity[N, U](v)
-
-    /** create a new unit Temperature of type U with numeric value of `this` */
-    def withTemperature[U](implicit t2k: DerivedTemp[U]): Temperature[N, U] =
-      new Temperature[N, U](v)
   }
 
   def coefficient[U1, U2](implicit cu: ConvertableUnits[U1, U2]): Rational = cu.coef
